+++ pages/operations.php
@@ -645,7 +645,7 @@ if ($_GET['action'] == 'importCategories') {
     if (move_uploaded_file($_FILES['jsonFile']['tmp_name'], $target_path)) {
         $data = json_decode(file_get_contents($target_path), true);
 
-        if (!$data['ingCategory'] && !$data['formulaCategories']) {
+        if (!$data['ingCategory'] && !$data['formulaCategories'] && !$data['ingProfiles']) {
             $result['error'] = "JSON File seems invalid. Please make sure you are importing the right file";
             echo json_encode($result);
             return;
@@ -675,6 +675,19 @@ if ($_GET['action'] == 'importCategories') {
                 if (!$stmt->execute()) {
                     $success = false;
                     $result['error'] = "Error inserting into formulaCategories: " . $stmt->error;
+                    break;
+                }
+            }
+            $stmt->close();
+        }
+		
+		if ($data['ingProfiles']) {
+            $stmt = $conn->prepare("INSERT INTO `ingProfiles` (`name`, `notes`, `image`) VALUES (?, ?, ?)");
+            foreach ($data['ingProfiles'] as $d) {
+                $stmt->bind_param("sss", $d['name'], $d['notes'], $d['image']);
+                if (!$stmt->execute()) {
+                    $success = false;
+                    $result['error'] = "Error inserting into ingProfiles: " . $stmt->error;
                     break;
                 }
             }
@@ -1000,4 +1013,38 @@ if($_GET['action'] == 'exportMaking'){
 	return;
 }
 
+//EXPORT INGREDIENT PROFILES
+if($_GET['action'] == 'exportIngProf'){
+	if(empty(mysqli_num_rows(mysqli_query($conn, "SELECT id FROM ingProfiles")))){
+		$msg['error'] = 'No data found to export.';
+		echo json_encode($msg);
+		return;
+	}
+	$data = 0;
+	$q = mysqli_query($conn, "SELECT * FROM ingProfiles");
+	while($resData = mysqli_fetch_assoc($q)){
+		
+		$r['id'] = (int)$resData['id'];
+		$r['name'] = (string)$resData['name']?: "-";
+		$r['notes'] = (string)$resData['notes']?: "-";
+		$r['image'] = (string)$resData['image'] ?: "-";
+		
+		$data++;
+		$cat[] = $r;
+	}
+	
+	$vd['product'] = $product;
+	$vd['version'] = $ver;
+	$vd['ingCategory'] = $data;
+	$vd['timestamp'] = date('d/m/Y H:i:s');
+
+	
+	$result['ingProfiles'] = $cat;
+	$result['pvMeta'] = $vd;
+	
+	header('Content-disposition: attachment; filename=IngProfiles.json');
+	header('Content-type: application/json');
+	echo json_encode($result, JSON_PRETTY_PRINT);
+	return;
+}
 ?>